Random Tree
There is a tree of nodes, each node contains a random integer number and
a block of memory of random size. The number of child nodes for any
parent nodes is random, and the depth of the tree is also a random
number. All the random numbers should be generated by random().

TASKS:
1. Create a tree of this kind.
2. Exam the tree, if a parent node and his child node both contain odd
numbers, then insert a node containing an even number between the parent
node and the child node. The inserted node may have only 1 child node.
3. Traverse the tree, and print the number of each node.
4. Delete the tree and free the memory.